Adjust API to the new catalog structure
