+++ saleor/graphql/utils.py
@@ -1,6 +1,4 @@
 import graphene
-from django.shortcuts import _get_queryset
-from django_prices.templatetags import prices_i18n
 
 
 class CategoryAncestorsCache:
@@ -27,35 +25,3 @@ class DjangoPkInterface(graphene.Interface):
 
     def resolve_pk(self, info):
         return self.pk
-
-
-class PriceType(graphene.ObjectType):
-    currency = graphene.String()
-    gross = graphene.Float()
-    gross_localized = graphene.String()
-    net = graphene.Float()
-    net_localized = graphene.String()
-
-    def resolve_gross_localized(self, info):
-        return prices_i18n.gross(self)
-
-    def resolve_net_localized(self, info):
-        return prices_i18n.net(self)
-
-
-class PriceField(graphene.Field):
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(lambda: PriceType, *args, **kwargs)
-
-
-class PriceRangeType(graphene.ObjectType):
-    max_price = PriceField()
-    min_price = PriceField()
-
-
-
-class PriceRangeField(graphene.Field):
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(lambda: PriceRangeType, *args, **kwargs)
